refactor: move auth tag presenter
